+++ VoxReader.UnitTests/RotationTests.cs
@@ -12,6 +12,8 @@ namespace VoxReader.UnitTests
         private const string RotationTestFile3 = "data/rotation_3.zip";
         private const string RotationTestFile4 = "data/rotation_4.zip";
 
+        // TODO: create tests for local/global position and local/global size
+        
         [Fact]
         public void TestRot()
         {